refactor: 공통으로 import 하도록 리팩토링 및 오류 수정
