+++ src/App.tsx
@@ -2,7 +2,6 @@ import Alert from "@components/commons/modal/Alert";
 import Confirm from "@components/commons/modal/Confirm";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
-import router from "@routes/Router";
 import GlobalStyle from "@styles/global";
 import theme from "@styles/theme";
 import { RouterProvider } from "react-router-dom";
@@ -12,6 +11,7 @@ import Modal from "@components/commons/modal/Modal";
 import { ThemeProvider as MuiThemeProvider, createTheme } from "@mui/material/styles";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+import router from "./routes/Router";
 
 const darkTheme = createTheme({
   palette: {
