+++ src/pages/book/components/paidBook/PaidBook.tsx
@@ -3,8 +3,7 @@ import { IconCheck } from "@assets/svgs";
 import Button from "@components/commons/button/Button";
 import Spacing from "@components/commons/spacing/Spacing";
 import Toast from "@components/commons/toast/Toast";
-import useModal from "@hooks/useModal";
-import useToast from "@hooks/useToast";
+import { useModal, useToast } from "@hooks";
 import { getBankNameKr } from "@utils/getBankName";
 import { getDeviceType } from "@utils/getDeviceType";
 import Lottie from "react-lottie-player";
