+++ src/pages/register/Register.tsx
@@ -12,12 +12,11 @@ import Spacing from "@components/commons/spacing/Spacing";
 import Stepper from "@components/commons/stepper/Stepper";
 import TimePicker from "@components/commons/timepicker/TimePicker";
 import { NAVIGATION_STATE } from "@constants/navigationState";
-import useLogin from "@hooks/useLogin";
-import useModal from "@hooks/useModal";
+import { useLogin, useModal } from "@hooks";
 import Content from "@pages/gig/components/content/Content";
 import ShowInfo from "@pages/gig/components/showInfo/ShowInfo";
 import { SHOW_TYPE_KEY } from "@pages/gig/constants";
-import { navigateAtom } from "@stores/navigate";
+import { navigateAtom } from "@stores";
 import { requestKakaoLogin } from "@utils/kakaoLogin";
 import { numericFilter, phoneNumberFilter, priceFilter } from "@utils/useInputFilter";
 import dayjs from "dayjs";
