+++ src/pages/register/RegisterComplete.tsx
@@ -1,12 +1,12 @@
-import Lottie from "react-lottie-player";
 import registerComplete from "@assets/lottie/register-complete.json";
-import * as S from "./Register.styled";
-import Spacing from "@components/commons/spacing/Spacing";
 import Button from "@components/commons/button/Button";
-import { useLocation, useNavigate } from "react-router-dom";
-import { useEffect, useState } from "react";
+import Spacing from "@components/commons/spacing/Spacing";
 import { NAVIGATION_STATE } from "@constants/navigationState";
-import { useHeader } from "@hooks/useHeader";
+import { useHeader } from "@hooks";
+import { useEffect } from "react";
+import Lottie from "react-lottie-player";
+import { useLocation, useNavigate } from "react-router-dom";
+import * as S from "./Register.styled";
 
 const RegisterComplete = () => {
   const user = "메이커";
