+++ src/pages/test/TestPage.tsx
@@ -8,7 +8,7 @@ import Spacing from "@components/commons/spacing/Spacing";
 import Stepper from "@components/commons/stepper/Stepper";
 import TimePicker from "@components/commons/timepicker/TimePicker";
 import Toast from "@components/commons/toast/Toast";
-import useToast from "@hooks/useToast";
+import { useToast } from "@hooks";
 import { nameFilter } from "@utils/useInputFilter";
 import { Dayjs } from "dayjs";
 import { ChangeEvent, useState } from "react";
