+++ src/pages/ticketholderlist/TicketHolderList.tsx
@@ -4,9 +4,7 @@ import Button from "@components/commons/button/Button";
 import Loading from "@components/commons/loading/Loading";
 import Toast from "@components/commons/toast/Toast";
 import { NAVIGATION_STATE } from "@constants/navigationState";
-import { useHeader } from "@hooks/useHeader";
-import useModal from "@hooks/useModal";
-import useToast from "@hooks/useToast";
+import { useHeader, useModal, useToast } from "@hooks";
 import { DeleteFormDataProps } from "@typings/deleteBookerFormatProps";
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
